fixed build errors remove old deprecated package
