--- packages/templates/typescript-multiple/.gitignore
@@ -1,5 +1,0 @@
-node_modules
-npm-debug.log
-dist
-temp
-yarn-error.log
