--- packages/templates/typescript-multiple/.npmignore
@@ -1,4 +1,0 @@
-src
-node_modules
-tests
-!dist
