--- packages/templates/typescript-multiple/.npmrc
@@ -1,1 +1,0 @@
-save-exact = true
