--- packages/templates/typescript-multiple/README.md
@@ -1,5 +1,0 @@
-# TypeScript Multiple Files Template
-
-This template generates TypeScript typings for both client side and server side.
-
-It very similar to the regular TypeScript template - [you can see here it's documentation](../typescript/README.md).
