--- packages/templates/typescript-multiple/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "graphql-codegen-typescript-template-multiple",
-  "version": "0.9.4",
-  "description": "",
-  "license": "MIT",
-  "scripts": {
-    "prepublishOnly": "yarn build",
-    "build": "codegen-templates-scripts build",
-    "pretest": "yarn build",
-    "test": "codegen-templates-scripts test"
-  },
-  "devDependencies": {
-    "@types/common-tags": "1.4.0",
-    "codegen-templates-scripts": "0.9.4",
-    "common-tags": "1.8.0",
-    "graphql-codegen-compiler": "0.9.4",
-    "graphql-codegen-core": "0.9.4"
-  },
-  "main": "./dist/index.js",
-  "typings": "dist/index.d.ts",
-  "typescript": {
-    "definition": "dist/index.d.ts"
-  },
-  "jest": {
-    "globals": {
-      "ts-jest": {
-        "enableTsDiagnostics": false
-      }
-    },
-    "transform": {
-      "^.+\\.tsx?$": "ts-jest"
-    },
-    "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "jsx",
-      "json",
-      "node"
-    ]
-  }
-}
