--- packages/templates/typescript-multiple/src/fragment.handlebars
@@ -1,19 +1,0 @@
-{{#eachImport this }}
-import { {{ name }} } from './{{ file }}';
-{{/eachImport}}
-export namespace {{name}} {
-  export type Fragment = {
-    {{#if @root.config.immutableTypes }}readonly {{/if}}__typename?: "{{ onType }}";
-    {{> selectionSet fields }}
-  }{{> fragments this }}
-  {{#each innerModels }}
-  export type {{ modelType }} = {
-  {{#unless hasInlineFragments}}
-    {{#if @root.config.immutableTypes }}readonly {{/if}}__typename?: "{{ schemaBaseType }}";
-  {{else}}
-    {{#if @root.config.immutableTypes }}readonly {{/if}}__typename?: {{#each inlineFragments}}{{onType}}["__typename"]{{#unless @last}} | {{/unless}}{{/each}};
-  {{/unless}}
-    {{> selectionSet fields }}
-  }{{> fragments this }}
-  {{/each}}
-}
