--- packages/templates/typescript-multiple/src/index.ts
@@ -1,3 +1,0 @@
-import { config } from './config';
-
-export default config;
