--- packages/templates/typescript-multiple/src/operation.handlebars
@@ -1,26 +1,0 @@
-{{#eachImport this }}
-import { {{ name }} } from './{{ file }}';
-{{/eachImport}}
-export namespace {{ toPascalCase name }} {
-  export type Variables = {
-{{#each variables}}
-    {{#if @root.config.immutableTypes }}readonly {{/if}}{{ name }}{{ getOptionals this }}: {{ convertedType this }};
-{{/each}}
-  }
-
-  export type {{ toPascalCase operationType }} ={{#if hasFields}} {
-    {{#if @root.config.immutableTypes }}readonly {{/if}}__typename?: "{{ toPascalCase operationType }}";
-    {{> selectionSet fields }}
-  }{{/if}}{{> fragments this }}
-  {{#each innerModels }}
-
-  export type {{ modelType }} ={{#if hasFields}} {
-  {{#unless hasInlineFragments}}
-    {{#if @root.config.immutableTypes }}readonly {{/if}}__typename?: "{{ schemaBaseType }}";
-  {{else}}
-    {{#if @root.config.immutableTypes }}readonly {{/if}}__typename?: {{#each inlineFragments}}{{onType}}["__typename"]{{#unless @last}} | {{/unless}}{{/each}};
-  {{/unless}}
-    {{> selectionSet fields }}
-  }{{/if}}{{> fragments this }}
-  {{/each}}
-}
