--- packages/templates/typescript-multiple/src/polyfills.d.ts
@@ -1,4 +1,0 @@
-declare module '*.handlebars' {
-  const content: string;
-  export = content;
-}
