--- packages/templates/typescript-multiple/src/scalar.handlebars
@@ -1,5 +1,0 @@
-{{#eachImport this }}
-import { {{ name }} } from './{{ file }}';
-{{/eachImport}}
-{{ toComment description }}
-export type {{ name }} = any;
