--- packages/templates/typescript-multiple/src/type.handlebars
@@ -1,19 +1,0 @@
-{{#eachImport this }}
-import { {{ name }} } from './{{ file }}';
-{{/eachImport}}
-{{ toComment description }}
-export interface {{ name }}{{#if hasInterfaces}} extends {{#each interfaces}}{{this}}{{#unless @last}},{{/unless}}{{/each}}{{/if}} {
-{{#each fields}}
-  {{#if @root.config.immutableTypes }}readonly {{/if}}{{ name }}{{ getOptionals this }}: {{ convertedType this }}; {{ toComment description }}
-{{/each}}
-}
-
-{{#each fields}}
-  {{~# if hasArguments }}
-export interface {{ toPascalCase name }}{{ toPascalCase ../name }}Args {
-{{#each arguments}}
-  {{#if @root.config.immutableTypes }}readonly {{/if}}{{ name }}{{ getOptionals this }}: {{ convertedType this }}; {{ toComment description }}
-{{/each}}
-}
-  {{/if}}
-{{/each}}
