--- packages/templates/typescript-multiple/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "experimentalDecorators": true,
-    "module": "commonjs",
-    "target": "es5",
-    "lib": ["es6", "esnext", "es2015"],
-    "noImplicitAny": false,
-    "suppressImplicitAnyIndexErrors": true,
-    "moduleResolution": "node",
-    "emitDecoratorMetadata": true,
-    "sourceMap": true,
-    "declaration": true,
-    "outDir": "./dist/"
-  },
-  "exclude": ["node_modules", "tests", "dist"]
-}
