+++ packages/templates/typescript/tests/typescript.spec.ts
@@ -933,154 +933,4 @@ describe('TypeScript template', () => {
       `);
     });
   });
-
-  describe('Resolvers', () => {
-    it('should contain the Resolver type', async () => {
-      const { context } = compileAndBuildContext(`
-        type Query {
-          fieldTest: String 
-        }
-        
-        schema {
-          query: Query
-        }
-      `);
-
-      const compiled = await compileTemplate(config, context);
-
-      const content = compiled[0].content;
-
-      expect(content).toBeSimilarStringTo(`
-        import { GraphQLResolveInfo } from 'graphql';
-
-        type Resolver<Result, Args = any> = (
-          parent: any,
-          args: Args,
-          context: any,
-          info: GraphQLResolveInfo
-        ) => Promise<Result> | Result;
-      `);
-    });
-
-    it('should make fields optional', async () => {
-      const { context } = compileAndBuildContext(`
-        type Query {
-          fieldTest: String 
-        }
-        
-        schema {
-          query: Query
-        }
-      `);
-
-      const compiled = await compileTemplate(config, context);
-
-      const content = compiled[0].content;
-
-      expect(content).toBeSimilarStringTo(`
-        export namespace QueryResolvers {
-          export interface Resolvers {
-            fieldTest?: FieldTestResolver;
-          }
-        `);
-    });
-
-    it('should provide a generic type of result', async () => {
-      const { context } = compileAndBuildContext(`
-        type Query {
-          fieldTest: String 
-        }
-        
-        schema {
-          query: Query
-        }
-      `);
-
-      const compiled = await compileTemplate(config, context);
-
-      const content = compiled[0].content;
-
-      expect(content).toBeSimilarStringTo(`
-        export namespace QueryResolvers {
-          export interface Resolvers {
-            fieldTest?: FieldTestResolver;
-          }
-    
-          export type FieldTestResolver = Resolver<string | null>;
-        }
-      `);
-    });
-
-    it('should provide a generic type of arguments and support optionals', async () => {
-      const { context } = compileAndBuildContext(`
-        type Query {
-          fieldTest(last: Int!, sort: String): String
-        }
-        
-        schema {
-          query: Query
-        }
-      `);
-
-      const compiled = await compileTemplate(config, context);
-
-      const content = compiled[0].content;
-
-      expect(content).toBeSimilarStringTo(`
-        export namespace QueryResolvers {
-          export interface Resolvers {
-            fieldTest?: FieldTestResolver;
-          }
-    
-          export type FieldTestResolver = Resolver<string | null, FieldTestArgs>;
-          
-          export interface FieldTestArgs {
-            last: number;
-            sort?: string | null;
-          }
-        }
-      `);
-    });
-
-    it('should handle resolvers flag, true by default', async () => {
-      const { context } = compileAndBuildContext(`
-        type Query {
-          fieldTest: String 
-        }
-        
-        schema {
-          query: Query
-        }
-      `);
-
-      const compiled = await compileTemplate(
-        {
-          ...config,
-          config: {
-            resolvers: false
-          }
-        },
-        context
-      );
-
-      const content = compiled[0].content;
-
-      expect(content).not.toBeSimilarStringTo(`
-        import { GraphQLResolveInfo } from 'graphql';
-      `);
-
-      expect(content).not.toBeSimilarStringTo(`
-        type Resolver<Result, Args = any> = (
-          parent: any,
-          args: Args,
-          context: any,
-          info: GraphQLResolveInfo
-        ) => Promise<Result> | Result;
-      `);
-
-      expect(content).not.toBeSimilarStringTo(`
-        export namespace QueryResolvers {
-      `);
-    });
-  });
 });
